+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ## Unreleased
 
-* The legacy "Lizmap HTML" popup is now deprecated for vector layer. A warning is raised when saving a the CFG file.
+* The legacy "Lizmap HTML" popup is now deprecated for vector layer. A warning is raised when saving the CFG file.
+* Fix the "OK" button was not closing the dialog.
 
 ## 3.9.6 - 2022-12-05